eebc-advisor/backend/requirements.txt
flask
flask-cors
pdfplumber
pypdf
sentence-transformers
faiss-cpu
pydantic
rapidfuzz
numpy
groq
gunicorn
